Navigation failed debug
BaseNavigationContainer.tsx:417 The action 'NAVIGATE' with payload {"name":"IndividualChatScreen","params":{"username":"Homelander"}} was not handled by any navigator.

Do you have a screen named 'IndividualChatScreen'?

If you're trying to navigate to a screen in a nested navigator, see https://reactnavigation.org/docs/nesting-navigators#navigating-to-a-screen-in-a-nested-navigator.

This is a development-only warning and won't be shown in production.